% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_flair_data.R
\name{delete_flair_data}
\alias{delete_flair_data}
\title{Delete FLAIR Image Filenames}
\usage{
delete_flair_data(...)
}
\arguments{
\item{...}{arguments to pass to \code{\link{delete_kirby21_data}},
\code{modalities = "FLAIR"} so it cannot be specified}
}
\value{
Nothing is returned
}
\description{
Delete the files for the FLAIR images
}
