% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_flair_data.R
\name{download_flair_data}
\alias{download_flair_data}
\title{Download FLAIR Image Filenames}
\usage{
download_flair_data(...)
}
\arguments{
\item{...}{arguments to pass to \code{\link{download_kirby21_data}},
\code{modalities = "FLAIR"} so it cannot be specified}
}
\value{
Indicator if the data is downloaded.
}
\description{
Download the files for the FLAIR images
}
