% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_flair_filenames.R
\name{get_flair_filenames}
\alias{get_flair_filenames}
\title{Get FLAIR Image Filenames}
\usage{
get_flair_filenames(...)
}
\arguments{
\item{...}{arguments to pass to \code{\link{get_image_filenames}},
\code{modalities = "FLAIR"} so it cannot be specified}
}
\value{
Vector fo filenames
}
\description{
Return the filenames for the FLAIR images
}
\examples{
get_flair_filenames()
}

